ASKFLOOR-1-T5b: hold file ceilings, extract group-processor deps from runtime-app, cap its bootstrap imports (run 9 + window)
Telegram callback-handlers back to 900, ipc-interaction-processing to 865,
session-commands to 740. The group-processor dependency module is a
bootstrap composition file; the architecture map classes all of app/ as
runtime, so it carries the same capped exception as its sibling wiring
files (maxViolations 3). Signals S-0090/S-0091 resolved.

Ticket: Q-0177-6001
