+++ .factory/signals.jsonl
@@ -176,3 +176,7 @@
 {"event": "resolved", "id": "S-0088-0a27", "at": "2026-09-08T13:54:06+00:00", "notes": "Scope widened: apps/core/src/channels/teams/message-actions.ts (readTeamsMessageAction decoder + handleTeamsMessageAction dispatch) added to the T5a write scope; 64 files, budget 68/6800 unchanged. Continue."}
 {"event": "raised", "id": "S-0089-3034", "task": "ASKFLOOR-1", "at": "2026-09-08T15:13:21+00:00", "generated_by": "implementer", "kind": "contradiction", "message": "T5a requires permission-card-affordances.ts to own the persisted host-derived model, but check:architecture forbids both application and runtime from importing that channels path; the write scope has no application/domain formatter module or architecture-exception file, so the required path, dependency direction, and green gate cannot all be satisfied without a contract resolution."}
 {"event": "resolved", "id": "S-0089-3034", "at": "2026-09-08T15:14:31+00:00", "notes": "Layering fix: the PermissionCardAffordances TYPE lives in NEW domain/permission-card-affordances.ts (domain imports nothing above it, so types.ts can reference it); the host-side DERIVATION (buildPermissionCardAffordances, scalar/batch shape, parse) lives in NEW application/permissions/permission-card-affordances.ts; channels/permission-card-affordances.ts keeps ONLY labels, lines and templates and imports the model from domain/application. Both new files are now in the write scope (66 files, budget 68/6800). Move the code accordingly; check:architecture must pass. Continue."}
+{"event": "raised", "id": "S-0090-8568", "task": "ASKFLOOR-1", "at": "2026-09-09T13:23:32+00:00", "generated_by": "implementer", "kind": "contradiction", "message": "ASKFLOOR-1-T5b: renaming app/bootstrap/runtime-group-processor-deps.ts to group-processor-deps.ts leaves it classified as runtime because architecture-map.json maps all apps/core/src/app to runtime; the required bootstrap layer-map change is outside write scope."}
+{"event": "resolved", "id": "S-0090-8568", "at": "2026-09-09T13:25:55+00:00", "notes": "Ruling: no architecture-map change. Use the established seam: add ONE entry to scripts/architecture-exceptions.json for apps/core/src/app/bootstrap/group-processor-deps.ts with rule forbidden_import_by_layer, reason 'group-processor deps compose Postgres runtime-store and config for group processing \u2014 same capped bootstrap-wiring seam as the sibling register files.', removeByPhase canonical-boundary-cleanup-phase, maxViolations 3 (exact current count). That file is added to the task scope by the orchestrator at stage done (amend-scope). Keep the group-processor-deps.ts rename. Then npm run check:architecture and npx tsc --noEmit must both pass; run the telegram, session-commands and runtime-app suites once; finish."}
+{"event": "raised", "id": "S-0091-ba8c", "task": "ASKFLOOR-1", "at": "2026-09-09T13:28:46+00:00", "generated_by": "implementer", "kind": "contradiction", "message": "ASKFLOOR-1-T5b requires app/bootstrap/group-processor-deps.ts to import config and Postgres adapters, but scripts/architecture-map.json classifies all apps/core/src/app paths as runtime; the prescribed rename cannot make npm run check:architecture pass, and the required map correction is outside write scope."}
+{"event": "resolved", "id": "S-0091-ba8c", "at": "2026-09-09T13:30:27+00:00", "notes": "Resolved by the orchestrator: scripts/architecture-exceptions.json is outside the companion write guard, so the one capped exception entry for app/bootstrap/group-processor-deps.ts is added host-side inside a ledgered degraded window (bounded coordinator fix, one file), recorded via amend-scope at stage done. Your run 9 product edits stand."}
