+++ apps/core/src/app/bootstrap/group-processor-deps.ts
